Fixed the message tests correctly.  The waypoint actor was working as the actor that should not show up in the GM.  I changed it to the crash actor, which is what it was supposed to be in the beginning.
Also cleaned up some size_t/unsigned int mess.  A size_t is only an unsigned int on windows, and on should not assume what it is.

+++ trunk/delta3d/tests/dtGame/gmloggertests.cpp
@@ -982,7 +982,7 @@ void GMLoggerTests::TestPlaybackRecordCycle()
       dtCore::System::GetInstance().Step();
 
       CPPUNIT_ASSERT_EQUAL_MESSAGE("Wrong number of game actors before changing to record.",
-         (unsigned)1,mGameManager->GetNumGameActors());
+         size_t(1),mGameManager->GetNumGameActors());
 
       //First, we'll start recording and then add some actors thus generating
       //actor create messages.
@@ -999,7 +999,7 @@ void GMLoggerTests::TestPlaybackRecordCycle()
       SLEEP(10);
       dtCore::System::GetInstance().Step();
       CPPUNIT_ASSERT_EQUAL_MESSAGE("Wrong number of actors correctly added during record.",
-         (unsigned)3,mGameManager->GetNumGameActors());
+         size_t(3),mGameManager->GetNumGameActors());
 
       testSignal->Reset();
       logController->RequestServerGetStatus();
